refactor(contracts): own the code index freshness DTOs
`tracedecay-code-index-runtime` depended on `tracedecay-dashboard-api`
only for the `code_index_freshness_api` wire types its scheduler
registry returns. Move the DTOs (`CodeIndexWorktreeFreshnessV1`,
`CodeIndexBuildProgressV1`, `CodeGraphServingReadinessV1`, the reader
port aliases, `CodeIndexFreshnessPayloadV1`, and their serde tests) to
`tracedecay_contracts::code_index_freshness`; the dashboard route keeps
only the HTTP handler and imports the types from there.

Serde/schemars output is unchanged: the exported dashboard contract
schema is byte-identical, so `dashboard/src/contracts` needs no
regeneration. Drops the dashboard-api dependency edge from the code
index runtime; consumer edits (cli, mcp, root) are import retargets.
