+++ crates/tracedecay-code-index-runtime/src/code_index_scheduler/tests/retained_configuration_tests.rs
@@ -131,7 +131,7 @@ async fn partitioned_restart_rebuilds_incompatible_retained_generation() {
         .expect("replacement generation");
     assert_eq!(
         latest.code_graph_serving_readiness(),
-        tracedecay_dashboard_api::code_index_freshness_api::CodeGraphServingReadinessV1::Ready,
+        tracedecay_contracts::code_index_freshness::CodeGraphServingReadinessV1::Ready,
         "the replacement must pass through graph activation before it seats"
     );
     assert!(
