+++ crates/tracedecay-mcp/src/handlers/info/status.rs
@@ -288,7 +288,7 @@ pub async fn handle_status(
                 && matches!(
                     freshness.code_graph_serving,
                     Some(
-                        tracedecay_dashboard_api::code_index_freshness_api::CodeGraphServingReadinessV1::Ready
+                        tracedecay_contracts::code_index_freshness::CodeGraphServingReadinessV1::Ready
                     )
                 )
         })
@@ -418,7 +418,7 @@ pub async fn handle_status(
 /// the warning carries the exact reason and remediation instead of a
 /// wait-longer message.
 fn code_index_freshness_projection(
-    freshness: &tracedecay_dashboard_api::code_index_freshness_api::CodeIndexWorktreeFreshnessV1,
+    freshness: &tracedecay_contracts::code_index_freshness::CodeIndexWorktreeFreshnessV1,
 ) -> (&'static str, Option<String>) {
     let authoritative = freshness.latest_generation_id.is_some()
         && freshness.coverage == "complete"
@@ -717,23 +717,22 @@ mod tests {
 
     #[test]
     fn a_parked_deterministic_violation_reports_parked_not_warming() {
-        let freshness =
-            tracedecay_dashboard_api::code_index_freshness_api::CodeIndexWorktreeFreshnessV1 {
-                worktree_root: "/project".to_owned(),
-                staleness_state: Some("parked".to_owned()),
-                coverage: "complete".to_owned(),
-                parked: Some(
-                    tracedecay_dashboard_api::code_index_freshness_api::CodeIndexConvergenceParkedV1 {
-                        reason: "code text artifacts root is not owner-private (mode 775, need 700)"
-                            .to_owned(),
-                        remediation: "restore owner-only access".to_owned(),
-                        parked_at_micros: 42,
-                        observed_passes: 3,
-                        retries_on_wake: true,
-                    },
-                ),
-                ..Default::default()
-            };
+        let freshness = tracedecay_contracts::code_index_freshness::CodeIndexWorktreeFreshnessV1 {
+            worktree_root: "/project".to_owned(),
+            staleness_state: Some("parked".to_owned()),
+            coverage: "complete".to_owned(),
+            parked: Some(
+                tracedecay_contracts::code_index_freshness::CodeIndexConvergenceParkedV1 {
+                    reason: "code text artifacts root is not owner-private (mode 775, need 700)"
+                        .to_owned(),
+                    remediation: "restore owner-only access".to_owned(),
+                    parked_at_micros: 42,
+                    observed_passes: 3,
+                    retries_on_wake: true,
+                },
+            ),
+            ..Default::default()
+        };
 
         let (status, warning) = code_index_freshness_projection(&freshness);
 
@@ -745,14 +744,13 @@ mod tests {
 
     #[test]
     fn status_freshness_preserves_typed_graph_serving_readiness() {
-        let freshness =
-            tracedecay_dashboard_api::code_index_freshness_api::CodeIndexWorktreeFreshnessV1 {
-                worktree_root: "/project".to_owned(),
-                code_graph_serving: Some(
-                    tracedecay_dashboard_api::code_index_freshness_api::CodeGraphServingReadinessV1::Ready,
-                ),
-                ..Default::default()
-            };
+        let freshness = tracedecay_contracts::code_index_freshness::CodeIndexWorktreeFreshnessV1 {
+            worktree_root: "/project".to_owned(),
+            code_graph_serving: Some(
+                tracedecay_contracts::code_index_freshness::CodeGraphServingReadinessV1::Ready,
+            ),
+            ..Default::default()
+        };
 
         let value = serde_json::to_value(freshness).expect("freshness serializes");
         assert_eq!(
@@ -763,23 +761,22 @@ mod tests {
 
     #[test]
     fn a_serving_worktree_with_a_parked_newer_build_stays_current_but_warns() {
-        let freshness =
-            tracedecay_dashboard_api::code_index_freshness_api::CodeIndexWorktreeFreshnessV1 {
-                worktree_root: "/project".to_owned(),
-                latest_generation_id: Some("generation.fixture".to_owned()),
-                staleness_state: Some("fresh".to_owned()),
-                coverage: "complete".to_owned(),
-                parked: Some(
-                    tracedecay_dashboard_api::code_index_freshness_api::CodeIndexConvergenceParkedV1 {
-                        reason: "code text artifacts root is not owner-private".to_owned(),
-                        remediation: "restore owner-only access".to_owned(),
-                        parked_at_micros: 42,
-                        observed_passes: 1,
-                        retries_on_wake: true,
-                    },
-                ),
-                ..Default::default()
-            };
+        let freshness = tracedecay_contracts::code_index_freshness::CodeIndexWorktreeFreshnessV1 {
+            worktree_root: "/project".to_owned(),
+            latest_generation_id: Some("generation.fixture".to_owned()),
+            staleness_state: Some("fresh".to_owned()),
+            coverage: "complete".to_owned(),
+            parked: Some(
+                tracedecay_contracts::code_index_freshness::CodeIndexConvergenceParkedV1 {
+                    reason: "code text artifacts root is not owner-private".to_owned(),
+                    remediation: "restore owner-only access".to_owned(),
+                    parked_at_micros: 42,
+                    observed_passes: 1,
+                    retries_on_wake: true,
+                },
+            ),
+            ..Default::default()
+        };
 
         let (status, warning) = code_index_freshness_projection(&freshness);
 
@@ -789,13 +786,12 @@ mod tests {
 
     #[test]
     fn an_unparked_incomplete_read_stays_warming() {
-        let freshness =
-            tracedecay_dashboard_api::code_index_freshness_api::CodeIndexWorktreeFreshnessV1 {
-                worktree_root: "/project".to_owned(),
-                staleness_state: Some("indexing".to_owned()),
-                coverage: "complete".to_owned(),
-                ..Default::default()
-            };
+        let freshness = tracedecay_contracts::code_index_freshness::CodeIndexWorktreeFreshnessV1 {
+            worktree_root: "/project".to_owned(),
+            staleness_state: Some("indexing".to_owned()),
+            coverage: "complete".to_owned(),
+            ..Default::default()
+        };
 
         let (status, warning) = code_index_freshness_projection(&freshness);
 
@@ -809,14 +805,13 @@ mod tests {
 
     #[test]
     fn a_ready_generation_under_source_verification_is_stale_not_warming() {
-        let freshness =
-            tracedecay_dashboard_api::code_index_freshness_api::CodeIndexWorktreeFreshnessV1 {
-                worktree_root: "/project".to_owned(),
-                latest_generation_id: Some("generation.fixture".to_owned()),
-                staleness_state: Some("verifying".to_owned()),
-                coverage: "partial_source_verification".to_owned(),
-                ..Default::default()
-            };
+        let freshness = tracedecay_contracts::code_index_freshness::CodeIndexWorktreeFreshnessV1 {
+            worktree_root: "/project".to_owned(),
+            latest_generation_id: Some("generation.fixture".to_owned()),
+            staleness_state: Some("verifying".to_owned()),
+            coverage: "partial_source_verification".to_owned(),
+            ..Default::default()
+        };
 
         let (status, warning) = code_index_freshness_projection(&freshness);
